[TASK] refactored utility namespace
