+++ Web/typo3conf/ext/sessions/Tests/Functional/Planning/UtilityTest.php
@@ -18,13 +18,13 @@ use TYPO3\CMS\Core\Tests\FunctionalTestCase;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Extbase\Domain\Model\FrontendUser;
 use TYPO3\Sessions\Domain\Model\AnySession;
-use TYPO3\Sessions\PlanningUtility;
+use TYPO3\Sessions\Planning\Utility;
 
 /**
- * Class PlanningUtilityTest
+ * Class UtilityTest
  * @package T3DD\Sessions\Tests\Unit
  */
-class PlanningUtilityTest extends FunctionalTestCase
+class UtilityTest extends FunctionalTestCase
 {
     /**
      * @var array
@@ -34,7 +34,7 @@ class PlanningUtilityTest extends FunctionalTestCase
     );
 
     /**
-     * @var PlanningUtility
+     * @var Utility
      */
     protected $subject;
 
@@ -45,16 +45,16 @@ class PlanningUtilityTest extends FunctionalTestCase
     {
         parent::setUp();
 
-        $this->importDataSet(__DIR__ . '/Fixtures/Xml/fe_users.xml');
-        $this->importDataSet(__DIR__ . '/Fixtures/Xml/tx_sessions_domain_model_session.xml');
-        $this->importDataSet(__DIR__ . '/Fixtures/Xml/tx_sessions_session_record_mm.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/Xml/fe_users.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/Xml/tx_sessions_domain_model_session.xml');
+        $this->importDataSet(__DIR__ . '/../Fixtures/Xml/tx_sessions_session_record_mm.xml');
 
         ExtensionManagementUtility::addTypoScriptSetup('
             <INCLUDE_TYPOSCRIPT: source="FILE:EXT:sessions/Configuration/TypoScript/setup.txt">
             config.tx_extbase.persistence < plugin.tx_sessions.persistence
         ');
 
-        $this->subject = $this->getAccessibleMock(PlanningUtility::class, array('_dummy'), array());
+        $this->subject = $this->getAccessibleMock(Utility::class, array('_dummy'), array());
     }
 
     /**
